refactor: nominee message component ui
