+++ packages/ui/src/components/atoms/index.ts
@@ -33,3 +33,4 @@ export * from './MainAppStyles';
 export * from './Tooltip';
 export * from './Video';
 export * from './AdvanceTextDiaplay';
+export * from './NomineeMessage';
